src/test/kotlin/canon/parser/xml/strategy/SubmitStrategyTest.kt
package canon.parser.xml.strategy

import canon.extension.toNode
import canon.parser.xml.CanonXmlParser
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Assertions.assertNotNull
import org.junit.jupiter.api.Test

class SubmitStrategyTest {

    @Test
    fun testParse() {
        val xml = "<submit id='testId' class='testClass' name='testName'>text</submit>"
        val parsed = SubmitStrategy().parse(xml.toNode(), HashMap<String, Any?>(), CanonXmlParser()::toRenderables)

        assertNotNull(parsed)
        assertEquals("testId", parsed.id)
        assertEquals("testClass", parsed.`class`)
        assertEquals("text", parsed.text)
        assertEquals("testName", parsed.name)
    }

}
